Address pull request command - Seperating CCacheEntry and putting in its own file - Removing exra line
